--- tedge/src/cli/connect/commands/connect_c8y.rs
@@ -1,121 +1,0 @@
-use crate::cli::connect::*;
-use crate::command::{Command, ExecutionContext};
-use crate::config::ConfigError;
-use mqtt_client::{Client, Message, Topic};
-use std::time::Duration;
-use tedge_config::*;
-use tokio::time::timeout;
-
-pub struct ConnectC8yCommand {
-    pub config_repository: TEdgeConfigRepository,
-}
-
-impl Command for ConnectC8yCommand {
-    fn description(&self) -> String {
-        "create bridge to connect Cumulocity cloud.".into()
-    }
-
-    fn execute(&self, context: &ExecutionContext) -> Result<(), anyhow::Error> {
-        let mut config = self.config_repository.load()?;
-        assign_default(&mut config, C8yRootCertPathSetting)?;
-        let bridge_config = create_c8y_bridge_config(&config)?;
-        self.config_repository.store(config)?;
-
-        bridge_config.new_bridge(&context.user_manager)?;
-        println!(
-            "Sending packets to check connection. This may take up to {} seconds.\n",
-            WAIT_FOR_CHECK_SECONDS
-        );
-        Ok(check_connection()?)
-    }
-}
-
-const RESPONSE_TIMEOUT: Duration = Duration::from_secs(10);
-const C8Y_CONFIG_FILENAME: &str = "c8y-bridge.conf";
-
-fn create_c8y_bridge_config(config: &TEdgeConfig) -> Result<BridgeConfig, ConfigError> {
-    let params = BridgeConfigParams {
-        connect_url: config.query(C8yUrlSetting)?,
-        mqtt_tls_port: MQTT_TLS_PORT,
-        config_file: C8Y_CONFIG_FILENAME.into(),
-        bridge_root_cert_path: config.query(C8yRootCertPathSetting)?,
-        remote_clientid: config.query(DeviceIdSetting)?,
-        bridge_certfile: config.query(DeviceCertPathSetting)?,
-        bridge_keyfile: config.query(DeviceKeyPathSetting)?,
-    };
-
-    Ok(BridgeConfig::new_for_c8y(params))
-}
-
-// Check the connection by using the response of the SmartREST template 100.
-// If getting the response '41,100,Device already existing', the connection is established.
-//
-// If the device is already registered, it can finish in the first try.
-// If the device is new, the device is going to be registered here and
-// the check can finish in the second try as there is no error response in the first try.
-
-#[tokio::main]
-async fn check_connection() -> Result<(), ConnectError> {
-    const C8Y_TOPIC_BUILTIN_MESSAGE_UPSTREAM: &str = "c8y/s/us";
-    const C8Y_TOPIC_ERROR_MESSAGE_DOWNSTREAM: &str = "c8y/s/e";
-    const CLIENT_ID: &str = "check_connection_c8y";
-
-    let c8y_msg_pub_topic = Topic::new(C8Y_TOPIC_BUILTIN_MESSAGE_UPSTREAM)?;
-    let c8y_error_sub_topic = Topic::new(C8Y_TOPIC_ERROR_MESSAGE_DOWNSTREAM)?;
-
-    let mqtt = Client::connect(CLIENT_ID, &mqtt_client::Config::default()).await?;
-    let mut error_response = mqtt.subscribe(c8y_error_sub_topic.filter()).await?;
-
-    let (sender, mut receiver) = tokio::sync::oneshot::channel();
-
-    let _task_handle = tokio::spawn(async move {
-        while let Some(message) = error_response.next().await {
-            if std::str::from_utf8(&message.payload)
-                .unwrap_or("")
-                .contains("41,100,Device already existing")
-            {
-                let _ = sender.send(true);
-                break;
-            }
-        }
-    });
-
-    for i in 0..2 {
-        print!("Try {} / 2: Sending a message to Cumulocity. ", i + 1,);
-
-        // 100: Device creation
-        mqtt.publish(Message::new(&c8y_msg_pub_topic, "100"))
-            .await?;
-
-        let fut = timeout(RESPONSE_TIMEOUT, &mut receiver);
-        match fut.await {
-            Ok(Ok(true)) => {
-                println!("Received expected response message, connection check is successful.\n",);
-                return Ok(());
-            }
-            _err => {
-                if i == 0 {
-                    println!("... No response. If the device is new, it's normal to get no response in the first try.");
-                } else {
-                    println!("... No response. ");
-                }
-            }
-        }
-    }
-
-    println!("Warning: Bridge has been configured, but Cumulocity connection check failed.\n",);
-    Ok(())
-}
-
-// XXX: Improve naming
-fn assign_default<T: ConfigSetting + Copy>(
-    config: &mut TEdgeConfig,
-    setting: T,
-) -> Result<(), ConfigError>
-where
-    TEdgeConfig: ConfigSettingAccessor<T>,
-{
-    let value = config.query(setting)?;
-    let () = config.update(setting, value)?;
-    Ok(())
-}
