--- tedge/src/cli/connect/commands/mod.rs
@@ -1,4 +1,0 @@
-mod connect_azure;
-mod connect_c8y;
-
-pub use self::{connect_azure::*, connect_c8y::*};
