+++ tedge/src/cli/connect/mod.rs
@@ -1,17 +1,17 @@
-pub use self::{bridge_config::*, cli::*, commands::*, common_mosquitto_config::*, error::*};
+pub use self::{
+    bridge_config::*, bridge_config_azure::*, bridge_config_c8y::*, cli::*, command::*,
+    common_mosquitto_config::*, error::*,
+};
 
 mod bridge_config;
-mod bridge_config_ext;
+mod bridge_config_azure;
+mod bridge_config_c8y;
 mod cli;
-mod commands;
+mod command;
 mod common_mosquitto_config;
 mod error;
-
-#[cfg(test)]
-mod test;
 
 const COMMON_MOSQUITTO_CONFIG_FILENAME: &str = "tedge-mosquitto.conf";
 const MOSQUITTO_RESTART_TIMEOUT_SECONDS: u64 = 5;
 const MQTT_TLS_PORT: u16 = 8883;
 const TEDGE_BRIDGE_CONF_DIR_PATH: &str = "mosquitto-conf";
-const WAIT_FOR_CHECK_SECONDS: u64 = 10;
